single source of fetch-polyfill
