+++ src/document-fetcher.js
@@ -15,13 +15,13 @@
  */
 
 import {FetchInitDef, setupInit} from './service/xhr-impl';
-import {Services} from '../../../src/services';
-import {dev, user} from './log';
-
+import {Services} from './services';
+import {fetchPolyfill} from './fetch-polyfill.js';
 
 /** @private @enum {number} Allowed fetch responses. */
 const allowedFetchTypes_ = {
   document: 1,
+  text: 2,
 };
 
 
@@ -31,11 +31,13 @@ export class DocumentFetcher {
    * @param {!Window} win
    */
   constructor(win) {
+    /** @type {!Window} */
     this.win_ = win;
 
+    /** @const @private {!./service/xhr-impl.Xhr} */
     this.xhr_ = Services.xhrFor(this.win_);
 
-    this.xhr_.setFetchOverride(this.fetchDocument_);
+    this.xhr_.setFetchOverride(fetchPolyfill);
   }
 
   /**
@@ -43,103 +45,12 @@ export class DocumentFetcher {
    *
    * @param {string} input
    * @param {?FetchInitDef=} opt_init
-   * @return {FetchResponse}
+   * @return {Response}
    */
   fetchDocument(input, opt_init) {
     const init = setupInit(opt_init, 'text/html');
     init.responseType = 'document';
-    return this.fetch(input, init)
+    return this.xhr_.fetch(input, init)
         .then(response => response.document_());
-  }
-
-  /**
-   * A minimal polyfill of Fetch API. It only polyfills what we currently use.
-   *
-   * See https://developer.mozilla.org/en-US/docs/Web/API/GlobalFetch/fetch
-   *
-   * Notice that the "fetch" method itself is not exported as that would require
-   * us to immediately support a much wide API.
-   *
-   * @param {Request|string} input
-   * @param {!FetchInitDef} init
-   * @return {!Promise<!Response>}
-   * @private
-   */
-  fetchDocument_(input, init) {
-    dev().assert(typeof input == 'string', 'Only URL supported: %s', input);
-    return new Promise((resolve, reject) => {
-      const xhr = this.createXhrRequest_(init.method || 'GET', input);
-
-      if (init.credentials == 'include') {
-        xhr.withCredentials = true;
-      }
-
-      if (('responseType' in init) &&
-        !(init.responseType in allowedFetchTypes_)) {
-        throw new Error('Document fetcher should only be used for responseType=document, please use Services.xhr for any other use case.');
-      } else {
-        xhr.responseType = 'document';
-      }
-
-      if (init.headers) {
-        Object.keys(init.headers).forEach(function(header) {
-          xhr.setRequestHeader(header, init.headers[header]);
-        });
-      }
-
-      xhr.onreadystatechange = () => {
-        if (xhr.readyState < /* STATUS_RECEIVED */ 2) {
-          return;
-        }
-        if (xhr.status < 100 || xhr.status > 599) {
-          xhr.onreadystatechange = null;
-          reject(user().createExpectedError(`Unknown HTTP status ${xhr.status}`));
-          return;
-        }
-
-        // TODO(dvoytenko): This is currently simplified: we will wait for the
-        // whole document loading to complete. This is fine for the use cases
-        // we have now, but may need to be reimplemented later.
-        if (xhr.readyState == /* COMPLETE */ 4) {
-          user().assert(xhr.responseXML,
-              'responseXML should exist. Make sure to return ' +
-              'Content-Type: text/html header.');
-          const doc = /** @type {!Document} */(dev().assert(xhr.responseXML));
-          return resolve(doc);
-        }
-      };
-      xhr.onerror = () => {
-        reject(user().createExpectedError('Network failure'));
-      };
-      xhr.onabort = () => {
-        reject(user().createExpectedError('Request aborted'));
-      };
-
-      if (init.method == 'POST') {
-        xhr.send(init.body);
-      } else {
-        xhr.send();
-      }
-    });
-  }
-
-  /**
-   * @param {string} method
-   * @param {string} url
-   * @return {!XMLHttpRequest|!XDomainRequest}
-   * @private
-   */
-  createXhrRequest_(method, url) {
-    let xhr = new XMLHttpRequest();
-    if ('withCredentials' in xhr) {
-      xhr.open(method, url, true);
-    } else if (typeof XDomainRequest != 'undefined') {
-      // IE-specific object.
-      xhr = new XDomainRequest();
-      xhr.open(method, url);
-    } else {
-      throw dev().createExpectedError('CORS is not supported');
-    }
-    return xhr;
   }
 }
